feat: move files about entity, repository interface, usecases to domain module
