--- common/src/main/kotlin/kr/ohyung/common/marks/Entity.kt
@@ -1,3 +1,0 @@
-package kr.ohyung.common.marks
-
-interface Entity
