--- common/src/main/kotlin/kr/ohyung/common/marks/Repository.kt
@@ -1,3 +1,0 @@
-package kr.ohyung.common.marks
-
-interface Repository
