+++ domain/src/main/kotlin/kr/ohyung/domain/Entity.kt
@@ -1,0 +1,3 @@
+package kr.ohyung.domain
+
+interface Entity
