+++ domain/src/main/kotlin/kr/ohyung/domain/Repository.kt
@@ -1,0 +1,3 @@
+package kr.ohyung.domain
+
+interface Repository
