+++ domain/src/main/kotlin/kr/ohyung/domain/UseCase.kt
@@ -1,0 +1,3 @@
+package kr.ohyung.domain
+
+interface UseCase
