+++ domain/src/main/kotlin/kr/ohyung/domain/exception/UserNotFoundException.kt
@@ -1,0 +1,3 @@
+package kr.ohyung.domain.exception
+
+class UserNotFoundException(message: String) : Exception(message)
