+++ domain/src/main/kotlin/kr/ohyung/domain/usecase/BaseUseCase.kt
@@ -1,0 +1,12 @@
+package kr.ohyung.domain.usecase
+
+import kr.ohyung.domain.UseCase
+
+/**
+ * Created by Lee Oh Hyung on 2020/08/23.
+ */
+abstract class BaseUseCase<in Params> : UseCase {
+
+    abstract fun execute(params: Params): Any
+
+}
