--- presentation/src/main/kotlin/kr/ohyung/navigation/data/repository/UserRepository.kt
@@ -1,10 +1,0 @@
-package kr.ohyung.navigation.data.repository
-
-import io.reactivex.Single
-import kr.ohyung.common.marks.Repository
-import kr.ohyung.navigation.data.entity.UserProfile
-
-interface UserRepository : Repository {
-
-    fun getUserProfile(userName: String): Single<UserProfile>
-}
