+++ presentation/src/main/kotlin/kr/ohyung/navigation/di/RepositoryModule.kt
@@ -6,7 +6,7 @@ import dagger.hilt.InstallIn
 import dagger.hilt.android.components.ActivityRetainedComponent
 import dagger.hilt.android.scopes.ActivityRetainedScoped
 import kr.ohyung.navigation.data.UserApi
-import kr.ohyung.navigation.data.repository.UserRepository
+import kr.ohyung.domain.repository.UserRepository
 import kr.ohyung.navigation.data.repository.UserRepositoryImpl
 
 @Module
